[dawn][native] Spanify Compute/Fragment/Vertex/State
All of these structs need to be spanified together as they all contain
lists of overridable constants, but this also requires spanifying
FragmentState::targets and VertexState::buffers at the same time.

Make webgpu_structs.h include IntegerTypes.h, which requires the
latter's dependency on dawn_platform.h.

Bug: 515272358, 526537224
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/320519
Commit-Queue: Corentin Wallez <cwallez@chromium.org>

+++ src/dawn/native/TintUtils.cpp
@@ -146,7 +146,7 @@ tint::VertexPullingConfig BuildVertexPullingTransformConfig(
     tint::VertexPullingConfig cfg;
     cfg.pulling_group = uint32_t(pullingBufferBindingSet);
 
-    cfg.vertex_state.resize(renderPipeline.GetVertexBufferCount());
+    cfg.vertex_state.resize(uint32_t(renderPipeline.GetVertexBufferCount()));
     for (VertexBufferSlot slot : renderPipeline.GetVertexBuffersUsed()) {
         const VertexBufferInfo& dawnInfo = renderPipeline.GetVertexBuffer(slot);
         tint::VertexBufferLayoutDescriptor* tintInfo =
